+++ HM/Tests/Files/FunctionFileNameUnitTest/plugins/regular-plugin/plugin.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace HM\Coding\Standards\Regular_Plugin;
+
+function foo() {}