+++ HM/Tests/Files/NamespaceDirectoryNameUnitTest/plugins/regular-plugin/plugin.php
@@ -1,0 +1,3 @@
+<?php
+
+namespace HM\Coding\Standards\Regular_Plugin;